Remove .travis.yml and rename test directory
