--- .travis.yml
@@ -1,12 +1,0 @@
-language: python
-dist: xenial
-
-python:
-- "3.7"
-
-install:
-- pip install cython pipenv
-- pipenv install --system --dev
-
-script:
-- python -m unittest discover tests
