+++ pyproject.toml
@@ -14,6 +14,7 @@ deprecation = "^2.0"
 [tool.poetry.dev-dependencies]
 black = {version = "^18.3-alpha.0", allows-prereleases = true}
 sphinx = "^2.2"
+parameterized = "^0.7.1"
 
 [build-system]
 requires = ["poetry>=0.12"]
