+++ test/test_song_effects.py
